requirements.txt
selenium
requests

python-dotenv
google-genai
openai